feat(ui): activity card component
