+++ src/pages/Activity/ActivityCard/interface.ts
@@ -1,0 +1,5 @@
+import type { Activity } from '../ActivityList/interface';
+
+export interface ActivityCardProps { 
+  content: Activity
+}
